Refactory bindings
* Update monero-cpp submodule
